% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_add_experiment.R
\name{cg_add_experiment}
\alias{cg_add_experiment}
\title{Add experimental data to a chronogram}
\usage{
cg_add_experiment(cg, experiment)
}
\arguments{
\item{cg}{a chronogram object (\verb{class cg_tbl})}

\item{experiment}{a \code{tibble::tibble()} containing experimental data. The dates
and IDs columns in \code{cg} must be present in \code{experiment}.}
}
\value{
A chronogram
}
\description{
Add experimental data to a chronogram
}
\examples{
## Example 1: A small study ##-------------------------------------
data(built_smallstudy)

## Setup ##
cg <- built_smallstudy$chronogram

infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe"
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)
## Add this new experiment data ##
cg_added <- cg_add_experiment(cg, infections_to_add)

## Example 2: Incorrect column names ##----------------------------

## Setup as for Example 1 ##
infections_to_add_renamed <- infections_to_add \%>\%
dplyr::rename(ID = elig_study_id)

## this fails ##
try(
cg_added_2 <- cg_add_experiment(cg, 
infections_to_add_renamed)
)

##------------------------------------------------------------------
}
\seealso{
\code{\link[=cg_assemble]{cg_assemble()}},
\code{\link[=cg_add_treatment]{cg_add_treatment()}},
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}},
\code{\link[=chronogram]{chronogram()}}
}
