% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_antigenic_history.R
\name{cg_annotate_antigenic_history}
\alias{cg_annotate_antigenic_history}
\title{Calculate and assign flags based on infection history}
\usage{
cg_annotate_antigenic_history(
  cg,
  episode_number = episode_number,
  dose_number = dose_number,
  episode_variant_summarised = episode_variant_summarised,
  ag_col = antigenic_history
)
}
\arguments{
\item{cg}{a chronogram}

\item{episode_number}{a character vector to identify the
episode_number column. Default is "episode_number".}

\item{dose_number}{a character vector to identify the column containing
the number of doses received on that day. Default is "dose_number".}

\item{episode_variant_summarised}{a character vector to identify
the column summarising the variant call. Default is
"episode_variant_summarised".}

\item{ag_col}{column with infection flag}
}
\value{
A chronogram
}
\description{
Calculate and assign flags based on infection history
}
\examples{
\dontrun{
cg <- cg_annotate_label_episodes(cg)
}

}
