% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_episodes_count.R
\name{cg_annotate_episodes_count}
\alias{cg_annotate_episodes_count}
\title{Calculate and assign flags based on infection history}
\usage{
cg_annotate_episodes_count(
  cg,
  episode_number = episode_number,
  count_col = count_col
)
}
\arguments{
\item{cg}{a chronogram}

\item{episode_number}{a character vector to use to label the ID
column. Default is "episode_number".}

\item{count_col}{column with infection count}
}
\value{
A subsetted chronogram
}
\description{
Count the total number of infections for each individual at the end
of the study (or at that data freeze), and make it available as a
fresh column. \code{cg_annotate_episodes_count()} should be run after
\code{cg_annotate_episodes_find()}. For the cumulative number of
infections use \code{cg_annotate_episodes_find()} only.
}
\examples{
library(dplyr)

data("built_smallstudy")
cg <- built_smallstudy$chronogram

## add infections to chronogram
cg <- cg_add_experiment(
  cg,
  built_smallstudy$infections_to_add
)

## annotate infections 
cg <- cg_annotate_episodes_find(
   cg,
   infection_cols = c("LFT", "PCR", "symptoms"),
   infection_present = c("pos", "Post", "^severe")
)

## count infections
cg <- cg_annotate_episodes_count(
cg)

## pull number of infections each individual has during study ##
cg \%>\% 
  dplyr::group_by(elig_study_id) \%>\% 
  dplyr::slice_head() \%>\% 
  dplyr::select(elig_study_id, count_col)

}
\seealso{
\code{\link[=cg_annotate_episodes_find]{cg_annotate_episodes_find()}}
}
