% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_episodes_fill.R
\name{cg_annotate_episodes_fill}
\alias{cg_annotate_episodes_fill}
\title{Fill experimental, treatment or symptom information within an infection episode}
\usage{
cg_annotate_episodes_fill(
  cg,
  col_to_fill,
  col_to_return,
  .direction = "down",
  episode_numbers_col = episode_number
)
}
\arguments{
\item{cg}{a chronogram}

\item{col_to_fill}{the name of the column to fill within each
episode}

\item{col_to_return}{the name of the returned, filled column}

\item{.direction}{See \code{tidyr::fill()}, provide quoted.}

\item{episode_numbers_col}{The column name to use for episode
numbers. Default is \code{episode_number}, unquoted.}
}
\value{
a chronogram, with episode data filled within each
episode.
}
\description{
Fill experimental, treatment or symptom information within an infection episode
}
\examples{
library(dplyr)

data("built_smallstudy")
cg <- built_smallstudy$chronogram

## add infections to chronogram ##
cg <- cg_add_experiment(
  cg,
  built_smallstudy$infections_to_add
)

## annotate infections ## 
cg <- cg_annotate_episodes_find(
   cg,
   infection_cols = c("LFT", "PCR", "symptoms"),
   infection_present = c("pos", "Post", "^severe")
)

## assign variants ##
cg <- cg \%>\%
mutate(
episode_variant =
case_when(
    # "is an episode" & "PCR positive" -> Delta #
    (!is.na(episode_number)) & PCR == "Pos" ~ "Delta",
    # "is an episode" & "PCR unavailable" -> Anc/Delta #
    (!is.na(episode_number)) & PCR == "not tested" ~ "Anc/Alpha"
)
)
## ^ this gives a variant call on a SINGLE row of each episode

## fill the variant call ##
cg <- cg \%>\% cg_annotate_episodes_fill(
 col_to_fill = episode_variant,
 col_to_return = episode_variant_filled,
 .direction = "updown",
 episode_numbers_col = episode_number
 )

}
