% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_episodes_find.R
\name{cg_annotate_episodes_find}
\alias{cg_annotate_episodes_find}
\title{Find infections per individual}
\usage{
cg_annotate_episodes_find(
  cg,
  infection_cols,
  infection_present,
  episode_days = 14,
  episode_numbers_col = episode_number,
  episode_start_col = episode_start,
  episode_end_col = episode_end
)
}
\arguments{
\item{cg}{a chronogram}

\item{infection_cols}{A vector of column names that contain
information regarding infection status, such as symptoms, LFT
results, PCR results, or sequencing results.}

\item{infection_present}{Strings used in each of \code{infection_cols}
to indicate the presence of infection.}

\item{episode_days}{The number of days to scan forwards and
backwards in time to consider events a single episode. For
example, is symptom onset \verb{10d} after a positive PCR test a
single episode? The default is \verb{14d}.}

\item{episode_numbers_col}{The column name to use for episode
numbers. Default is \code{episode_number} (unquoted).}

\item{episode_start_col, episode_end_col}{The column names to store
the episode start and end dates (defaults: \code{episode_start} and
\code{episode_end}, both unquoted).}
}
\value{
x a chronogram, with episode numbers annotated
}
\description{
Find infections per individual
}
\examples{
## Example 1: A small study ##-------------------------------------
data(built_smallstudy)
cg_small <- built_smallstudy$chronogram
infections_to_add <- built_smallstudy$infections_to_add

## add infection data to to chronogram ##
cg_small <- cg_add_experiment(cg_small, infections_to_add)

## now infection finding ##
cg_small_inf <- cg_annotate_episodes_find(cg_small,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "^severe")
)

summary(cg_small_inf$episode_number)

## exact text matching ##------------------------------------------
test2 <- cg_annotate_episodes_find(cg_small_inf,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("Pos", "Post", "^mild")
)

summary(test2$episode_number)

## empty strings will error (as they otherwise match everything) ##
test3a <- 
try(
cg_annotate_episodes_find(cg_small_inf,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "")
)
)
## a 'random' string will not error ##-----------------------------
test3b <- try(
cg_annotate_episodes_find(cg_small_inf,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "a")
)
)
}
\seealso{
\code{\link[=cg_annotate_episodes_find_seroconversion]{cg_annotate_episodes_find_seroconversion()}}
}
