% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_vaccines_count.R
\name{cg_annotate_vaccines_count}
\alias{cg_annotate_vaccines_count}
\title{Calculate and assign flags based on vaccine history}
\usage{
cg_annotate_vaccines_count(
  cg,
  vaccine_date_stem = NULL,
  dose = dose,
  dose_counter = NULL,
  intermediate_days = 14
)
}
\arguments{
\item{cg}{a chronogram}

\item{vaccine_date_stem}{the start of the vaccine date columns.
Default is NULL. Try "date_dose"; asssumes "date_dose_{1,2,3...}"}

\item{dose}{a character vector to use to label the dose column.
This is the numerical dose, filled only on the date of that dose
(and empty on intervening days). Default is "dose".}

\item{dose_counter}{column to return the cumulative number of
doses. Default is NULL. Try "dose_number"}

\item{intermediate_days}{number of days to pad after dose date to
return as eg "1star" if within intermediate days after dose 1,
numeric}
}
\value{
An annotated chronogram (class cg_tbl).
}
\description{
Calculate and assign flags based on vaccine history
}
\examples{
data("built_smallstudy")
cg <- built_smallstudy$chronogram

## annotate vaccines ##
cg <- cg_annotate_vaccines_count(cg,
  dose = dose,
  dose_counter = dose_number,
  vaccine_date_stem = date_dose,
  intermediate_days = 7
)

}
