% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_assemble.R
\name{cg_assemble}
\alias{cg_assemble}
\title{Assemble a chronogram}
\usage{
cg_assemble(
  start_date,
  end_date,
  calendar_date_col,
  metadata,
  metadata_ids_col,
  experiment_data_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{the start date, in a format recognised by
\code{lubridate::dmy()}}

\item{end_date}{the end date, in a format recognised by
\code{lubridate::dmy()}}

\item{calendar_date_col}{user-defined column name for dates
(unquoted).}

\item{metadata}{a tibble containing metadata}

\item{metadata_ids_col}{the (unquoted) column name for participant
IDs in metadata tibble. No default provided.}

\item{experiment_data_list}{a list of tibbles of experiment data.
See \code{cg_add_experiment()} for details. Ignored if not provided.}

\item{verbose}{Default TRUE. Show progress messages?}
}
\value{
a chronogram (\verb{class cg_tbl})
}
\description{
\code{cg_assemble()} assembles a chronogram from 4 pieces of
input data:
\itemize{
\item start date
\item end date
\item metadata
\item and an optional list of experiment data.
}

A verbose messsaging option (on by default) is provided to help
troubleshoot input.

Extra experimental can be added later with \code{cg_add_experiment()}.

For fine grain troubleshooting, run \code{chronogram_skeleton()} and
\code{chronogram()} sequentially (these are called in turn by
\code{cg_assemble()}). In all other circumstances, \code{cg_assemble()} is
the encouraged method.
}
\examples{
## Example 1: A small study ##-------------------------------------
data(smallstudy)

## Setup ##
start <- "01012020"
end <- "10102021"
meta <- smallstudy$small_study_metadata ## age, sex, vaccine dates
ab <- smallstudy$small_study_Ab ## antibody response data

## Assembly ##
cg_small <- cg_assemble(

## start and end date ##
start_date = start,
end_date = end,

## metadata ##
metadata = meta,
metadata_ids_col = elig_study_id,

## experiment data ##
experiment_data_list = list(ab),

## set the date column name ##
calendar_date_col = calendar_date
)

## Example 2: now with 2 types of experimental data ##-------------

## Setup as example 1, and second experimental dataset here: ##
infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe"
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)

cg <- cg_assemble(
  start_date = "01012020",
  end_date = "10102021",
  metadata = smallstudy$small_study_metadata,
  metadata_ids_col = elig_study_id,
  calendar_date_col = calendar_date,
  experiment_data_list = list(ab, infections_to_add)
)
##------------------------------------------------------------------
}
\seealso{
\code{\link[=cg_add_experiment]{cg_add_experiment()}} \code{\link[=chronogram_skeleton]{chronogram_skeleton()}}
\code{\link[=chronogram]{chronogram()}} \code{vignette("SQL_assembly", package = "chronogram")}
}
