% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_plot.R
\name{cg_plot}
\alias{cg_plot}
\title{cg_plot}
\usage{
cg_plot(
  cg,
  x = NULL,
  y_values,
  drop_vars = TRUE,
  point_alpha = 0.4,
  point_shape = 20,
  link_obs = TRUE,
  link_colour = "grey",
  link_alpha = 0.4,
  ...
)
}
\arguments{
\item{cg}{chronogram}

\item{x}{a column of time to use as x axis. If NULL, will default
to the chronogram's calendar date attribute. A user may want to
derive and use alternatives eg 'daysSinceDose2'.}

\item{y_values}{column within chronogram containing the data you
wish to plot.}

\item{drop_vars}{Default TRUE. See description.}

\item{point_alpha}{alpha for \code{geom_point()}.}

\item{point_shape}{shape for \code{geom_point()}.}

\item{link_obs}{Default TRUE. Draw a line to link results from same
individual?}

\item{link_colour}{colour for \code{geom_line()}}

\item{link_alpha}{alpha for \code{geom_line()}}

\item{...}{passed to \code{aes()}}
}
\description{
Create a \code{ggplot2} object, plotting a
user-defined y axis over time.

\code{ggplot2} objects retain the entirety of the provided dataset.
This allows later adjustments, such as adding extra geom_layers
with new information, or applying facets. To find this data
examine \code{obj$data}. If you save \code{ggplot2} objects, all source
data is ALSO saved. \code{cg_plot()} removes any un-used data by
default (\code{drop_vars=TRUE}). In writing a study specific
\code{ggplot2}, it is best practice to select minimal columns before
calling \code{ggplot()}.
}
\examples{

library(ggplot2)
library(patchwork)

data("built_smallstudy")
cg <- built_smallstudy$chronogram

p1 <- cg_plot_meta(cg,
  visit = serum_Ab_S
)

p2 <- cg_plot(cg,
  y_values = serum_Ab_S
)

p2 / p1

(p2 + scale_y_log10()) / p1

}
