% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_plot_meta.R
\name{cg_plot_meta}
\alias{cg_plot_meta}
\title{cg_plot_meta}
\usage{
cg_plot_meta(
  cg,
  x = NULL,
  date_dose_1 = date_dose_1,
  dose_1 = dose_1,
  date_dose_2 = date_dose_2,
  dose_2 = dose_2,
  visit = visit,
  drop_vars = TRUE,
  fill = NULL
)
}
\arguments{
\item{cg}{chronogram}

\item{x}{a column of time to use as x axis. If NULL, will default
to the chronogram's calendar date attribute. A user may want to
derive and use alternatives eg 'daysSinceDose2'.}

\item{date_dose_1}{column containing the date of dose.}

\item{dose_1}{column containing the vaccine formulation.}

\item{date_dose_2}{column containing the date of dose.}

\item{dose_2}{column containing the vaccine formulation.}

\item{visit}{column within chronogram to indicate a study visit
(i.e. samples available; NA when samples not taken, !=NA when
samples available). In our small study example, serum_Ab_S fills
this brief.}

\item{drop_vars}{Default TRUE. See description.}

\item{fill}{column used to determine fill}
}
\description{
Function to create a \code{ggplot2} object, plotting a
user-defined metadata over time in a swimmer's plot.

\code{ggplot2} objects retain the entirety of the provided dataset.
This allows later adjustments, such as adding extra geom_layers
with new information, or applying facets. To find this data
examine \code{obj$data}. If you save \code{ggplot2} objects, all source
data is ALSO saved. \code{cg_plot_meta()} removes any un-used data by
default (\code{drop_vars=TRUE}). In writing a study specific
\code{ggplot2}, it is best practice to select minimal columns before
calling \code{ggplot()}.
}
\examples{
\dontrun{
data("built_smallstudy")
cg <- built_smallstudy$chronogram

p1 <- cg_plot_meta(cg,
  visit = serum_Ab_S
)

p2 <- cg_plot(cg,
  y_values = serum_Ab_S
)
library(patchwork)
p2 / p1

(p2 + scale_y_log10()) / p1
}

}
