% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_save.R
\name{cg_save}
\alias{cg_save}
\title{Save a chronogram}
\usage{
cg_save(cg, file = NULL, ..., verbose = TRUE)
}
\arguments{
\item{cg}{a chronogram object (\verb{class cg_tbl})}

\item{file}{quoted filename. Include an extension.}

\item{...}{passed to \code{\link[base:save]{base::save()}}. For example, \code{compress}}

\item{verbose}{logical. Default \code{TRUE}. Display messages?}
}
\value{
INVISIBLE, a saved file
}
\description{
Write a chronogram to disk without duplicating metadata.
}
\examples{
data("built_smallstudy")
cg_orig <- built_smallstudy$chronogram
ff <- tempfile()
cg_save(cg_orig, file = ff)
cg_new <- cg_load(ff)

summary(cg_new == cg_orig)
identical(cg_new, cg_orig)

}
\seealso{
\code{\link[=cg_load]{cg_load()}}
}
