% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_select.R
\name{cg_select}
\alias{cg_select}
\title{Adapted dplyr::select() to include calendar dates and ID columns}
\usage{
cg_select(cg, ...)
}
\arguments{
\item{cg}{a chronogram}

\item{...}{passed to \code{dplyr::select()}}
}
\value{
a chronogram with selected columns, and the two columns
containing calendar dates and participant IDs.
}
\description{
Adapted dplyr::select() to include calendar dates and ID columns
}
\examples{
library(dplyr)

data("built_smallstudy")

cg <- built_smallstudy$chronogram

## keep or drop columns using selection helpers ##-----------------
## here, keep column names containing "dose"
cg_select(cg, contains("dose"))

## Or, drop column names containing "dose"
cg_select(cg, ! contains("dose"))

## keep or drop columns using exact names ##-----------------------
cg_select(cg, "dose_1")
# or equivalently:
cg_select(cg, dose_1)

# or several selectors together ##---------------------------------
cg_select(cg, dose_1, dose_2)

cg_select(cg, contains("dose") & contains("date"))

}
