% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_select_visit.R
\name{cg_select_visit}
\alias{cg_select_visit}
\title{Select a visit within a windowed chronogram}
\usage{
cg_select_visit(cg, visit_col = NULL, visit = c("earliest", "latest"))
}
\arguments{
\item{cg}{a windowed chronogram}

\item{visit_col}{a character vector to use to label the column with
either dates of visits, or results. This is usually an
experimental data column. Default is "NULL".}

\item{visit}{either "earliest" or "latest"}
}
\value{
A subsetted chronogram
}
\description{
Select a visit within a windowed chronogram
}
\examples{
\dontrun{
SevenDaysPrePostDose3 <- cg_window_by_metadata(
  annotatedChronogram, "date_dose_3", 7, 7
)
SevenDaysPrePostDose3 <- cg_select_visit(
  SevenDaysPrePostDose3, "earliest"
)
}

}
