% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_window_by_episode.R
\name{cg_window_by_episode}
\alias{cg_window_by_episode}
\title{Pick a window of dates from a chronogram: with respect to episode
start.}
\usage{
cg_window_by_episode(
  cg,
  episode_numbers_col = "episode_number",
  episode_handling = c("first", "last", "all"),
  preceding_days,
  following_days
)
}
\arguments{
\item{cg}{a chronogram}

\item{episode_numbers_col}{The column name to use for episode
numbers. Default is \code{episode_number}.}

\item{episode_handling}{which episode to reference. Must be one of
"first", "last", or "all".}

\item{preceding_days}{used as \code{filter( date > (date_col - preceding_days) )}}

\item{following_days}{used as \code{filter( date < (date_col + following_days) )}}
}
\value{
A windowed chronogram
}
\description{
Pick a window of dates from a chronogram: with respect to episode
start.
}
\examples{
\dontrun{

SevenDaysPrePostFirstEpisode <- cg_window_by_episode(
  annotatedChronogram, "first", 7, 7
)
}

}
\seealso{
\code{\link[=cg_window_by_visit]{cg_window_by_visit()}},
\code{\link[=cg_window_by_metadata]{cg_window_by_metadata()}}
}
