% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_window_by_metadata.R
\name{cg_window_by_metadata}
\alias{cg_window_by_metadata}
\title{Take a window of dates from a chronogram}
\usage{
cg_window_by_metadata(cg, windowing_date_col, preceding_days, following_days)
}
\arguments{
\item{cg}{a chronogram}

\item{windowing_date_col}{the column containing reference date (must repeat
the date for each row of that individual's record).}

\item{preceding_days}{used as filter( date > (windowing_date_col -
preceding_days) )}

\item{following_days}{used as filter( date < (windowing_date_col +
following_days) )}
}
\value{
A windowed chronogram
}
\description{
Works with respect to a column (eg date of dose 3).
}
\examples{
\dontrun{
SevenDaysPrePostDose3 <- cg_window_by_metadata(
  annotatedChronogram, "date_dose_3", 7, 7
)
}

}
\seealso{
\code{\link[=cg_window_by_episode]{cg_window_by_episode()}},
\code{\link[=cg_window_by_visit]{cg_window_by_visit()}}
}
