% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_experiment.R
\name{check_experiment}
\alias{check_experiment}
\title{check ExperimentalData}
\usage{
check_experiment(x, cg = NULL)
}
\arguments{
\item{x}{a tibble containing ExperimentalData, with an ID column
and a date of SAMPLE (not date of assay).}

\item{cg}{the chronogram to join}
}
\value{
"valid" or stop with message
}
\description{
check ExperimentalData
}
