% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{check metadata}
\usage{
check_metadata(x, cg_skeleton = NULL)
}
\arguments{
\item{x}{a tibble containing metadata}

\item{cg_skeleton}{chronogram skeleton}
}
\value{
TRUE or stop with message
}
\description{
check metadata
}
