% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronogram.R
\name{chronogram}
\alias{chronogram}
\title{Build a chronogram, from a chronogram_skeleton and metadata}
\usage{
chronogram(cg_skeleton, metadata)
}
\arguments{
\item{cg_skeleton}{a cg_skeleton, generated by
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}}}

\item{metadata}{a tibble containing metadata and ID column name
(and IDs) that are present in cg_skeleton}
}
\value{
An object of chronogram class, with metadata spread across
all calendar_dates.
}
\description{
Decorate a chronogram skeleton with metadata to return a
chronogram. For most circumstances, \code{cg_assemble()} is suggested
(which calls \code{chronogram_skeleton()} and \code{chronogram()}). See
assembly vignette.
}
\examples{
## a 3-person chronogram_skeleton ##-------------------------------

small_study <- chronogram_skeleton(
  col_ids = elig_study_id,
  ids = c(1, 2, 3),
  start_date = "01012020",
  end_date = "10102021",
  col_calendar_date = calendar_date
)

## A tibble containing some metadata for our 3 individuals ##
data(smallstudy)
small_study_metadata <- smallstudy$small_study_metadata
small_study_metadata

## Make a chronogram ##
small_study_chronogram <- chronogram(
  small_study,
  small_study_metadata
)

}
\seealso{
\code{\link[=cg_assemble]{cg_assemble()}},
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}},
\code{\link[=cg_add_experiment]{cg_add_experiment()}}
}
