% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dplyr.R
\name{dplyr_reconstruct.cg_tbl}
\alias{dplyr_reconstruct.cg_tbl}
\title{Dplyr reconstruct for chronogram}
\usage{
\method{dplyr_reconstruct}{cg_tbl}(data, template)
}
\arguments{
\item{data}{data}

\item{template}{example class}
}
\value{
a cg_tbl after a dplyr verb action
}
\description{
As recommended here:
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
\details{
Ensures dplyr verbs return a cg_tbl, rather than tbl.
User can ignore.
}
