% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dplyr.R
\name{dplyr_reconstruct.grouped_cg_df}
\alias{dplyr_reconstruct.grouped_cg_df}
\title{Dplyr reconstruct for grouped chronogram}
\usage{
\method{dplyr_reconstruct}{grouped_cg_df}(data, template)
}
\arguments{
\item{data}{data}

\item{template}{example class}
}
\value{
a grouped_cg_df after a dplyr verb action
}
\description{
As recommended here:
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
\details{
Ensures dplyr verbs return a grouped_cg_df, rather than grouped_df.
User can ignore.
}
