% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dplyr.R
\name{dplyr_row_slice.cg_tbl}
\alias{dplyr_row_slice.cg_tbl}
\title{Dplyr row slice for chronogram}
\usage{
\method{dplyr_row_slice}{cg_tbl}(data, i, ...)
}
\arguments{
\item{data}{data}

\item{i}{rows to eg slice}

\item{...}{additional arguments}
}
\value{
a cg_tbl after a dplyr verb action
}
\description{
As recommended here:
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
\details{
Ensures dplyr verbs return a cg_tbl, rather than tbl.
User can ignore.
}
