% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dplyr.R
\name{dplyr_row_slice.grouped_cg_df}
\alias{dplyr_row_slice.grouped_cg_df}
\title{Dplyr row slice for grouped chronogram}
\usage{
\method{dplyr_row_slice}{grouped_cg_df}(data, i, ..., preserve = FALSE)
}
\arguments{
\item{data}{data}

\item{i}{rows to eg slice}

\item{...}{additional arguments}

\item{preserve}{FALSE}
}
\value{
a grouped_cg_df after a dplyr verb action
}
\description{
As recommended here:
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
\details{
Ensures dplyr verbs return a grouped_cg_df, rather than grouped_df.
User can ignore.
}
