% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_chronogram.R
\name{glimpse.cg_tbl}
\alias{glimpse.cg_tbl}
\title{Glimpse S3 method for chronogram objects (class cg_tbl)}
\usage{
\method{glimpse}{cg_tbl}(x, width = NULL, ...)
}
\arguments{
\item{x}{a chrongram}

\item{width, ...}{passed to pillar::glimpse}
}
\value{
print to console
}
\description{
Glimpse S3 method for chronogram objects (class cg_tbl)
}
\examples{
library(dplyr)
data("built_smallstudy")
cg <- built_smallstudy$chronogram
glimpse(cg)

}
