% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_chronogram.R
\name{glimpse_experiment_data}
\alias{glimpse_experiment_data}
\alias{glimpse_experiment_data.cg_tbl}
\title{Glimpse S3 method for experiment data from chronogram objects (class cg_tbl)}
\usage{
glimpse_experiment_data(x, width = NULL, ...)

\method{glimpse_experiment_data}{cg_tbl}(x, width = NULL, ...)
}
\arguments{
\item{x}{a chrongram}

\item{width, ...}{passed to pillar::glimpse}
}
\value{
print to console
}
\description{
Glimpse S3 method for experiment data from chronogram objects (class cg_tbl)
}
\examples{
\dontrun{
data("built_smallstudy")
cg <- built_smallstudy$chronogram
glimpse_experiment_data(cg)
}
}
\seealso{
\code{\link[=glimpse]{glimpse()}}
}
