% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dplyr.R
\name{group_data_trim}
\alias{group_data_trim}
\title{dplyr:::group_data_trim}
\usage{
group_data_trim(group_data)
}
\arguments{
\item{group_data}{grouped dataset}
}
\description{
dplyr:::group_data_trim
}
\references{
dplyr:::group_data_trim
}
