% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chronogram.R
\name{new_chronogram}
\alias{new_chronogram}
\title{new chronogram}
\usage{
new_chronogram(x, metadata_cols, ...)
}
\arguments{
\item{x}{a tbl, a daughter of a join between a cg_skl_tbl and
one (or more) objects coerced to a tbl.}

\item{metadata_cols}{the column names from the metadata tibble}

\item{...}{passed to \code{new_tibble()}. Adds grouping support.}
}
\value{
x an extended tibble with class cg_tbl
}
\description{
new chronogram
}
\examples{
\dontrun{
x <- new_chronogram(x)
}
}
\seealso{
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}}
}
