% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chronogram_skeleton.R
\name{new_chronogram_skeleton}
\alias{new_chronogram_skeleton}
\title{new_chronogram_skeleton}
\usage{
new_chronogram_skeleton(x, calendar_date_col = NULL, ids_col = NULL)
}
\arguments{
\item{x}{a tibble}

\item{calendar_date_col}{provided as
rlang::enquo(an_unquoted_string)}

\item{ids_col}{provided as rlang::enquo(an_unquoted_string)}
}
\value{
x a tibble with class cg_skl_tbl
}
\description{
A low level constructor function. Use the helper
chronogram_skeleton() instead.
}
\seealso{
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}}
}
