% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_grouped_chronogram.R
\name{new_grouped_chronogram}
\alias{new_grouped_chronogram}
\title{New grouped chronogram}
\usage{
new_grouped_chronogram(x, groups, ..., class = character())
}
\arguments{
\item{x}{a chronogram}

\item{groups}{groups}

\item{...}{parsed to \code{dplyr::group_by()}}

\item{class}{character vector of class}
}
\value{
an extended grouped tibble with class \code{grouped_cg_df}
}
\description{
A constructor for a grouped chronogram. This is called automatically
by the S3 method: \code{group_by.cg_tbl()}.
}
