% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitch_chronogram}
\alias{pitch_chronogram}
\title{PITCH study: pre-built chronogram}
\format{
\subsection{\code{pitch_chronogram} A chronogram}{

\describe{
\item{pitch_chronogram}{Example pitch_chronogram}
}
}
}
\usage{
pitch_chronogram
}
\description{
A provided example derived from the de-identified PITCH dataset
available here, under a CC-BY-4.0 licence:
https://data.mendeley.com/datasets/fyp26zjgmj/1
}
\details{
Published: 3 November 2021,

DOI: 10.17632/fyp26zjgmj.1

Contributors: Rebecca Payne, Susan Hopkins, Victoria Hall,
Christina Dold, Christopher Duncan, Alex Richter, Miles Carroll,
Gavin Screaton, Thushan de Silva, Lance Turtle, Paul Klenerman,
Susanna Dunachie, PITCH Consortium authors.

The dataset supports Payne et al. Immunogenicity of standard and
extended dosing intervals of BNT162b2 mRNA vaccine Cell 2021.

NOTE: Changes have been made to the de-identified public dataset
(DOI: 10.17632/fyp26zjgmj.1).
\enumerate{
\item the public data reports dates as
MM/YYYY. To build a chronogram we have assumed 15/MM/YYYY for dose
1, and used the available intervals in days to place the remaining
data in time. For illustration purposes, we need plausible
DD/MM/YYYY dates - they are not real.
\item not all of the public
dataset has been used in this example (some assays not included).
}

NB: If your motivation is to re-analyse Payne et al. for scientific
reasons, rather than to explore this package, do not use this
example dataset. Instead, consult the de-identified Mendeley data
(DOI: 10.17632/fyp26zjgmj.1), or the Cell manuscript's data and
code availability statement.
}
\examples{
data("pitch_chronogram")
}
\keyword{datasets}
