% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_chronogram_skeleton.R
\name{print.cg_skl_tbl}
\alias{print.cg_skl_tbl}
\title{print chronogram skeleton}
\usage{
\method{print}{cg_skl_tbl}(x, ...)
}
\arguments{
\item{x}{a chronogram skeleton object (class cg_skl_tbl)}

\item{...}{additional arguments passed to tibble::print S3 generic}
}
\value{
print to console
}
\description{
print chronogram skeleton
}
\examples{
\dontrun{
small.study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = "01012020",
  end_date = "10102021"
)

## print, with default tibble options ##
small.study

## print, with eg 3 rows ##
print(small.study, n = 3)
}
}
