% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_chronogram.R
\name{print.cg_tbl}
\alias{print.cg_tbl}
\title{print chronogram}
\usage{
\method{print}{cg_tbl}(x, ...)
}
\arguments{
\item{x}{a chronogram object (class tbl_chronogram)}

\item{...}{additional arguments passed to tibble print S3 generic}
}
\value{
print to console
}
\description{
print chronogram
}
\examples{
\dontrun{
## a 3-person chronogram_skeleton ##
small_study <- chronogram_skeleton(
  col_ids = elig_study_id,
  ids = c(1, 2, 3),
  start_date = "01012020",
  end_date = "10102021",
  col_calendar_date = calendar_date
)

## Create a tibble containing some metadata for our 3 individuals ##
small_study_metadata <- tibble::tribble(
  ~elig_study_id, ~age, ~sex, ~dose_1, ~date_dose_1, ~dose_2, ~date_dose_2,
  1, 40, "F", "AZD1222", "05/01/2021", "AZD1222", "05/02/2021",
  2, 45, "F", "BNT162b2", "05/01/2021", "BNT162b2", "05/02/2021",
  3, 35, "M", "BNT162b2", "10/01/2021", "BNT162b2", "10/03/2021"
)

## Set appropriate metadata column classes ##
library(dplyr)
small_study_metadata <- small_study_metadata \%>\%
  mutate(across(c(sex, dose_1, dose_2), ~ as.factor(.x)))

small_study_metadata <- small_study_metadata \%>\%
  mutate(across(contains("date"), ~ lubridate::dmy(.x)))

## Make a chronogram ##
small_study_chronogram <- chronogram(
  small_study,
  small_study_metadata
)

## print, with default tibble options ##
small_study_chronogram

## print, with eg 3 rows ##
print(small_study_chronogram, n = 3)
}
}
