% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallstudy}
\alias{smallstudy}
\title{Small study}
\format{
\subsection{\code{smallstudy} A list of three components}{

\describe{
\item{small_study_ids}{IDs used in
small study to identify these 3 individuals. A vector.}
\item{small_study_metadata}{Example metadata. A tibble.}
\item{small_study_Ab}{Example experimental data. A tibble.}
}
}
}
\usage{
smallstudy
}
\description{
A fictional covid-19 vaccine study of 3 individuals. Provided a
vector of participant IDs, a tibble of metadata, and a tibble of
experimental data.
}
\examples{
data("smallstudy")
}
\keyword{datasets}
