% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_grouped_chronogram.R
\name{ungroup.grouped_cg_df}
\alias{ungroup.grouped_cg_df}
\title{ungroup-by chronogram}
\usage{
\method{ungroup}{grouped_cg_df}(x, ...)
}
\arguments{
\item{x}{input chronogram, grouped_cg_df}

\item{..., }{arguments expected by group_by S3}
}
\description{
An S3 implementation of \code{dplyr::ungroup()} that accepts a
\code{grouped_cg_df} and returns a chronogram. This
retains the attributes from the orginal chronogram.
}
\seealso{
\code{\link[=group_by]{group_by()}}
}
