% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chronogram.R
\name{validate_chronogram}
\alias{validate_chronogram}
\title{Validate chronogram}
\usage{
validate_chronogram(x)
}
\arguments{
\item{x}{a chronogram object (class cg_tbl)}
}
\value{
TRUE if valid
}
\description{
Checks that:
}
\details{
\itemize{
\item class is \code{cg_tbl}
\item the date column stored in attributes is present in data
\item the ID column stored in attributes is present in data
\item each date x participant ID is present only once (or absent).
}

Rarely needed by the user. \code{cg_assemble()} and annotation functions
use this to check returned \code{cg} is valid.
}
\seealso{
\code{\link[=cg_assemble]{cg_assemble()}}
\code{vignette("chronogram_class", package = "chronogram")}
}
