% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chronogram_skeleton.R
\name{validate_chronogram_skeleton}
\alias{validate_chronogram_skeleton}
\title{validate chronogram skeleton}
\usage{
validate_chronogram_skeleton(x)
}
\arguments{
\item{x}{a chronogram skeleton object (class tbl_df, with attributes that record IDs and date columns)}
}
\value{
Errors, or TRUE if valid chronogram skeleton
}
\description{
validate chronogram skeleton
}
