% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{built_smallstudy}
\alias{built_smallstudy}
\title{Small study: pre-built chronogram}
\format{
\subsection{\code{built_smallstudy}}{

A list of 2 components
\describe{
\item{small_study_chronogram_skeleton}{Example chronogram_skeleton}
\item{small_study_chronogram}{Example chronogram}
}
}
}
\usage{
built_smallstudy
}
\description{
A fictional covid-19 vaccine study of 3 individuals. Provided a
vector of participant IDs, a tibble of metadata, and a tibble of
experimental data.
}
\details{
The data are identical to \code{smallstudy}. \code{cg_assemble()} has been
run to shorten examples.
}
\examples{
data("built_smallstudy")
}
\keyword{datasets}
