% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_add_treatment.R
\name{cg_add_treatment}
\alias{cg_add_treatment}
\title{A helper function to add treatment data to a chronogram}
\usage{
cg_add_treatment(cg, treatment)
}
\arguments{
\item{cg}{a chronogram object (class cg_tbl)}

\item{treatment}{a tibble containing treatment dates (with columns:
the specified calendar_date, and ID columns, defined with
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}})

This code wraps the\code{cg_add_experiment} function.}
}
\value{
A chronogram
}
\description{
A helper function to add treatment data to a chronogram
}
\examples{
\dontrun{

## Create a tibble of exemplar treatment data ##
# Here, we have treated ID=1 with B cell depletion
# therapy, rituximab (RTX).
#
small_study_treatment <-
  tibble::tribble(
    ~elig_study_id, ~calendar_date, ~treatment,
    1, "01/02/2021", "RTX"
  )
small_study_treatment <- small_study_treatment \%>\%
  mutate(across(contains("date"), ~ lubridate::dmy(.x)))

## Add to chronogram ##
small_study_chronogram <- cg_add_treatment(
  small_study_chronogram,
  small_study_treatment
)
}
}
\seealso{
\code{\link[=cg_add_experiment]{cg_add_experiment()}}
}
