% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_episodes_count.R
\name{cg_annotate_episodes_count}
\alias{cg_annotate_episodes_count}
\title{Calculate and assign flags based on infection history}
\usage{
cg_annotate_episodes_count(
  x,
  episode_number = episode_number,
  count_col = count_col
)
}
\arguments{
\item{x}{a chronogram}

\item{episode_number}{a character vector to use to label the ID
column. Default is "episode_number".}

\item{count_col}{column with infection count}
}
\value{
A subsetted chronogram
}
\description{
This counts the total number of infections for each individual at
the end of the study (or at that data freeze), and makes it
available as a fresh column. this function should be run after
\code{cg_annotate_episodes_find()}. For the cumulative number of
infections use \code{cg_annotate_episodes_find()} only.
}
\examples{
\dontrun{
cg <- cg_annotate_episodes_count(cg)
}

}
\seealso{
\code{\link[=cg_annotate_episodes_find]{cg_annotate_episodes_find()}}
}
