% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_annotate_episodes_find.R
\name{cg_annotate_episodes_find}
\alias{cg_annotate_episodes_find}
\title{Find infections per individual}
\usage{
cg_annotate_episodes_find(
  x,
  infection_cols,
  infection_present,
  episode_days = 14,
  episode_numbers_col = episode_number,
  episode_start_col = episode_start,
  episode_end_col = episode_end
)
}
\arguments{
\item{x}{a chronogram}

\item{infection_cols}{Columns that might contain information
regarding infection status, such as symptoms, LFT results, PCR
results, or sequencing results.}

\item{infection_present}{Strings used in each of \code{infection_cols}
to indicate the presence of infection.}

\item{episode_days}{The number of days within which to consider
events a single episode. For example, is symptom onset 10d after
a positive PCR test a single episode? Default is 14d.}

\item{episode_numbers_col}{The column name to use for episode
numbers. Default is \code{episode_number} (unquoted).}

\item{episode_start_col}{The column names to store the episode
start and end dates (defaults: episode_start and episode_end, both unquoted).}

\item{episode_end_col}{The column names to store the episode start
and end dates (defaults: episode_start and episode_end, both unquoted).}
}
\value{
x a chronogram, with episode numbers annotated
}
\description{
Find infections per individual
}
\examples{
\dontrun{
data(built_smallstudy)
small_study <- built_smallstudy$chronogram

## Simulate some infection data ##
infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe",
  "11042021", "2", "pos", NA, "mild",
  "11082021", "2", "pos", "Pos", "mild",
  "11112021", "2", "pos", "Pos", NA
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)
## add to chronogram
small_study <- cg_add_experiment(small_study, infections_to_add)

## now infection finding ##
small_study_inf <- cg_annotate_episodes_find(small_study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "^severe")
)
summary(small_study_inf$episode_number)

## exact text matching ##
test2 <- cg_annotate_episodes_find(small_study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("Pos", "Post", "^mild")
)
summary(test2$episode_number)

## empty strings will error (as they otherwise match everything) ##
test3a <- cg_annotate_episodes_find(small_study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "")
)
## a 'random' string will not error ##
test3b <- cg_annotate_episodes_find(small_study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "a")
)

summary(test2$episode_number)
}
}
\seealso{
\code{\link[=cg_annotate_episodes_find_seroconversion]{cg_annotate_episodes_find_seroconversion()}}
}
