% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_assemble.R
\name{cg_assemble}
\alias{cg_assemble}
\title{Wrapper to assemble chronogram}
\usage{
cg_assemble(
  start_date,
  end_date,
  calendar_date_col,
  metadata,
  metadata_ids_col,
  experiment_data_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{the start date, in a format recognised by
\code{lubridate::dmy()}}

\item{end_date}{the end date, in a format recognised by
\code{lubridate::dmy()}}

\item{calendar_date_col}{to label the date column  (unquoted).}

\item{metadata}{a tibble containing metadata and ID column name (and
IDs) that are present in cg_skeleton}

\item{metadata_ids_col}{the (unquoted) column name for IDs in
metadata tibble. Studies may use StudyID, StudyId, PID, etc. No
default provided.}

\item{experiment_data_list}{a list of tibbles of experiment data.
See \code{cg_add_experiment()} for details. Ignored if not provided.}

\item{verbose}{Default TRUE. Show progress messages?}
}
\value{
a chronogram (class cg_tbl)
}
\description{
\code{cg_assemble()} wraps \code{chronogram_skeleton()},
\code{chronogram()} and \code{cg_add_experiment()}. A verbose messsaging
option (on by default) is provided to help troubleshoot input. For
further troubleshooting, running \code{chronogram_skeleton()},
\code{chronogram()} and \code{cg_add_experiment()} sequentially is
suggested.
}
\examples{
\dontrun{
data("smallstudy")


cg <- cg_assemble(
  start_date = "01012020",
  end_date = "10012020",
  metadata = smallstudy$small_study_metadata,
  metadata_ids_col = elig_study_id,
  calendar_date_col = calendar_date
)

## Now with experiment data ##
small_study_Ab <-
  tibble::tribble(
    ~elig_study_id, ~calendar_date, ~serum_Ab_S, ~serum_Ab_N,
    1, "05/01/2021", 500, 100,
    1, "15/01/2021", 4000, 100,
    1, "03/02/2021", 3750, 100,
    1, "15/02/2021", 10000, 100,
    2, "05/01/2021", 0, 0,
    2, "15/01/2021", 4000 / 2, 0,
    2, "03/02/2021", 3750 / 2, 0,
    2, "15/02/2021", 10000 / 2, 0,
    3, "05/01/2021", 0, 0,
    3, "25/01/2021", 4000 / 2, 0,
    3, "03/02/2021", 3750 / 2, 0,
    3, "20/03/2021", 10000 / 2, 0
  )

small_study_Ab <- small_study_Ab \%>\%
  mutate(across(contains("date"), ~ lubridate::dmy(.x)))

cg <- cg_assemble(
  start_date = "01012020",
  end_date = "10012020",
  metadata = smallstudy$small_study_metadata,
  metadata_ids_col = elig_study_id,
  calendar_date_col = calendar_date,
  experiment_data_list = list(small_study_Ab)
)

## Or two items of experiment data ##
infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe"
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)

cg <- cg_assemble(
  start_date = "01012020",
  end_date = "10102021",
  metadata = smallstudy$small_study_metadata,
  metadata_ids_col = elig_study_id,
  calendar_date_col = calendar_date,
  experiment_data_list = list(small_study_Ab, infections_to_add)
)
}

}
\seealso{
\code{\link[=chronogram_skeleton]{chronogram_skeleton()}} \code{\link[=chronogram]{chronogram()}}
\code{\link[=cg_add_experiment]{cg_add_experiment()}}
}
