% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_select_visit.R
\name{cg_select_visit}
\alias{cg_select_visit}
\title{Select a visit within a windowed chronogram}
\usage{
cg_select_visit(cg, visit_col = NULL, visit = c("earliest", "latest"))
}
\arguments{
\item{cg}{a windowed chronogram}

\item{visit_col}{a character vector to use to label the column with
either dates of visits, or results. This is usually an
experimental data column. Default is "NULL".}

\item{visit}{either "earliest" or "latest"}
}
\value{
A subsetted chronogram
}
\description{
Select a visit within a windowed chronogram
}
\examples{
data(pitch_chronogram)

SevenDaysPrePostDose2 <- cg_window_by_metadata(
  pitch_chronogram, dose_2_date, 7, 7
)

SevenDaysPrePostDose2

## every visit has the assay `Cov_2S_MSD`
## therefore use this column to indicate a study visit 
SevenDaysPrePostDose2_visits <- cg_select_visit(
  SevenDaysPrePostDose2, Cov_2S_MSD, "earliest"
)

SevenDaysPrePostDose2_visits 

## all days without visits are now dropped. Compare:
nrow(SevenDaysPrePostDose2)
nrow(SevenDaysPrePostDose2_visits)

}
