% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronogram_skeleton.R
\name{chronogram_skeleton}
\alias{chronogram_skeleton}
\title{Build a chronogram skeleton, a two column index of IDs and calendar
dates, a scaffold for a chronogram}
\usage{
chronogram_skeleton(
  ids = NULL,
  col_ids = column_name_for_ids,
  start_date = NULL,
  end_date = NULL,
  col_calendar_date = column_name_for_calendar_date
)
}
\arguments{
\item{ids}{a vector of study, or participant IDs.}

\item{col_ids}{to label the ID
column. Default is \code{column_name_for_ids} (unquoted). Studies may use
StudyID, StudyId, PID, etc. User advised to change to suit.}

\item{start_date}{the start date, in a format recognised by
\code{lubridate::dmy()}}

\item{end_date}{the end date, in a format recognised by
\code{lubridate::dmy()}}

\item{col_calendar_date}{to label the date
column. Default is \code{column_name_for_calendar_date} (unquoted).
User advised to change to suit.}
}
\value{
A chronogram skeleton class
}
\description{
Build a chronogram skeleton, a two column index of IDs and calendar
dates, a scaffold for a chronogram
}
\examples{
\dontrun{
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = "01012020",
  end_date = "10102021"
)

## Duplicate dates, or dates lubridate::dmy() does not recognise ##
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = c("01012020", "05012020"),
  end_date = "10102021"
)
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = c("01012020"),
  end_date = c("10102021", "20102020")
)
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = "20200101",
  end_date = "10102021"
)
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3),
  start_date = "01012020",
  end_date = "20201010"
)

## Duplicate participant IDs are accepted, but give a warning ##
small_study <- chronogram_skeleton(
  ids = c(1, 2, 3, 3),
  start_date = "01012020",
  end_date = "10102021"
)
}
}
