% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_chronogram.R
\name{glimpse.grouped_cg_df}
\alias{glimpse.grouped_cg_df}
\title{Glimpse S3 method for grouped chronogram objects (class grouped_cg_df)}
\usage{
\method{glimpse}{grouped_cg_df}(x, width = NULL, ...)
}
\arguments{
\item{x}{a grouped chrongram}

\item{width, ...}{passed to pillar::glimpse}
}
\value{
print to console
}
\description{
Glimpse S3 method for grouped chronogram objects (class grouped_cg_df)
}
\examples{
library(dplyr)

data("built_smallstudy")
cg <- built_smallstudy$chronogram
cg <- cg \%>\% group_by(elig_study_id)
glimpse(cg)

}
