% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_grouped_chronogram.R
\name{group_by.cg_tbl}
\alias{group_by.cg_tbl}
\title{Group-by chronogram}
\usage{
\method{group_by}{cg_tbl}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))
}
\arguments{
\item{.data}{input chronogram, cg_tbl}

\item{..., .add, .drop}{arguments expected by group_by S3}
}
\description{
An S3 implementation of \code{dplyr::group_by()} that accepts a
chronogram and returns an object of \code{grouped_cg_df} class. This
retains the attributes from the orginal chronogram.
}
