% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chronogram.R
\name{validate_chronogram}
\alias{validate_chronogram}
\title{validate chronogram}
\usage{
validate_chronogram(x)
}
\arguments{
\item{x}{a chronogram object (class cg_tbl)}
}
\value{
TRUE if valid
}
\description{
validate chronogram
}
